+++ clients/trieve-shopify-extension/package.json
@@ -45,6 +45,7 @@
     "clsx": "^2.1.1",
     "date-fns": "4.1.0",
     "deepmerge": "4.3.1",
+    "hono": "^4.7.7",
     "isbot": "^5.1.0",
     "jsonwebtoken": "^9.0.2",
     "nanoid": "^5.1.5",
